+++ ErsatzTV.Application/ProgramSchedules/Commands/CopyProgramSchedule.cs
@@ -1,0 +1,6 @@
+﻿using ErsatzTV.Core;
+
+namespace ErsatzTV.Application.ProgramSchedules;
+
+public record CopyProgramSchedule
+    (int ProgramScheduleId, string Name) : IRequest<Either<BaseError, ProgramScheduleViewModel>>;