+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerBase.cs
@@ -461,10 +461,14 @@ public abstract class PlayoutModeSchedulerBase<T> : IPlayoutModeScheduler<T> whe
         foreach (FillerPreset padFiller in Optional(
                      allFiller.FirstOrDefault(f => f.FillerMode == FillerMode.Pad && f.PadToNearestMinute.HasValue)))
         {
-            var totalDuration =
-                TimeSpan.FromMilliseconds(
-                    result.Sum(pi => (pi.Finish - pi.Start).TotalMilliseconds) +
+            var totalDuration = TimeSpan.FromMilliseconds(result.Sum(pi => (pi.Finish - pi.Start).TotalMilliseconds));
+
+            // add primary content to totalDuration only if it hasn't already been added
+            if (result.All(pi => pi.MediaItemId != playoutItem.MediaItemId))
+            {
+                totalDuration += TimeSpan.FromMilliseconds(
                     effectiveChapters.Sum(c => (c.EndTime - c.StartTime).TotalMilliseconds));
+            }
 
             int currentMinute = (playoutItem.StartOffset + totalDuration).Minute;
             // ReSharper disable once PossibleInvalidOperationException