+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
 - Fix many QSV pipeline bugs
+- Fix MPEG2 video format with QSV and VAAPI acceleration
 
 ## [0.7.3-beta] - 2023-01-25
 ### Added