+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fix many QSV pipeline bugs
 - Fix MPEG2 video format with QSV and VAAPI acceleration
+- Fix playback of content with undefined colorspace
 
 ## [0.7.3-beta] - 2023-01-25
 ### Added