+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix playback of content with undefined colorspace
 - Fix NVIDIA color normalization with VP9 sources
 - Fix fallback filler looping
+- Fix bug where some libraries would never scan
 
 ### Changed
 - Merge generated `Other Video` folder tags with tags from sidecar NFO